refactor(state): extract isSshTransportFailure to ssh-transport module (#6188)
Address advisor PRA-6 and PRA-10:

- Move isSshTransportFailure from src/lib/state/sandbox.ts into a new
  src/lib/state/ssh-transport.ts. The function was already a pure
  utility with no sandbox.ts internals; extraction keeps sandbox.ts
  smaller and gives future SSH-transport helpers a natural home. Re-
  exported from sandbox.ts for backwards compatibility.

- Rename src/lib/state/sandbox.test.ts to
  src/lib/state/ssh-transport.test.ts so the test file name matches
  its actual scope.

No behavior change.
